backend/requirements.txt
requests==2.31.0
supabase==2.24.0
python-dotenv==1.0.0
modal==1.2.2
# Supabase dependencies (installed automatically but pinned for compatibility)
gotrue==2.12.4
httpx==0.28.1
websockets==15.0.1
httpcore==1.0.9
